requirements.txt
fastapi
uvicorn
pymupdf
httpx
python-dotenv
python-multipart
pymongo
motor